+++ src/components/common/BottomSheet.tsx
@@ -1,4 +1,5 @@
 import clsx from 'clsx';
+import { CloseIcon } from '../icons/Icons';
 
 interface BottomSheetProps {
   isOpen: boolean;
@@ -32,7 +33,7 @@ export const BottomSheet = ({ isOpen, onClose, children }: BottomSheetProps) => 
         <div className="flex justify-end p-4">
           {/* 닫기 버튼 */}
           <button type='button' className="text-gray-500 hover:text-gray-700 text-xl" onClick={onClose}>
-            ✕
+            <CloseIcon/>
           </button>
         </div>
         <div className="p-6">{children}</div>
